+++ src/main/java/solutions/fairdata/openrefine/metadata/storage/VirtuosoStorage.java
@@ -22,22 +22,55 @@
  */
 package solutions.fairdata.openrefine.metadata.storage;
 
+import com.google.common.collect.ImmutableList;
+import org.apache.xerces.impl.dv.util.Base64;
 import solutions.fairdata.openrefine.metadata.dto.storage.StorageDTO;
 
 import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Paths;
 import java.util.HashMap;
+import java.util.List;
 
 public class VirtuosoStorage extends Storage {
 
     public static final String TYPE = "virtuoso";
+    private static final Charset CHARSET = StandardCharsets.UTF_8;
+
+    private static final String DETAIL_HOST = "host";
+    private static final String DETAIL_USERNAME = "username";
+    private static final String DETAIL_PASSWORD = "password";
+    private static final String DETAIL_DIRECTORY = "directory";
+    public static final List<String> DETAILS = ImmutableList.of(
+            DETAIL_HOST,
+            DETAIL_USERNAME,
+            DETAIL_PASSWORD,
+            DETAIL_DIRECTORY
+    );
+
+    private String host;
+    private String username;
+    private String password;
+    private String directory;
+
+    private void loadDetails(StorageDTO storageDTO) {
+        host = storageDTO.getDetails().get(DETAIL_HOST);
+        username = storageDTO.getDetails().get(DETAIL_USERNAME);
+        password = storageDTO.getDetails().get(DETAIL_PASSWORD);
+        directory = storageDTO.getDetails().getOrDefault(DETAIL_DIRECTORY, "/");
+
+        if (host == null) {
+            throw new IllegalArgumentException("Missing host for Virtuoso storage");
+        }
+    }
 
     public VirtuosoStorage(StorageDTO storageDTO) {
         super(storageDTO);
+        loadDetails(storageDTO);
     }
 
     @Override
@@ -46,13 +79,18 @@ public class VirtuosoStorage extends Storage {
     }
 
     @Override
+    public List<String> getDetailNames() {
+        return DETAILS;
+    }
+
+    @Override
     public String getFilePath(HashMap<String, String> metadata) {
-        return Paths.get(storageDTO.getDirectory(), metadata.getOrDefault("filenameExt", "")).toString();
+        return Paths.get(directory, metadata.getOrDefault("filenameExt", "")).toString();
     }
 
     @Override
     public String getURL(HashMap<String, String> metadata) {
-        return "http://" + storageDTO.getHost() + getFilePath(metadata);
+        return "http://" + host + getFilePath(metadata);
     }
 
     @Override
@@ -64,7 +102,7 @@ public class VirtuosoStorage extends Storage {
         if (forbidsContentType(contentType)) {
             throw new IOException("Bad content type for selected storage");
         }
-        makeVirtuosoPut(getURL(metadata), new String(data, StandardCharsets.UTF_8), contentType + "; charset=\"UTF-8\"");
+        makeVirtuosoPut(getURL(metadata), new String(data, CHARSET), contentType + "; charset=\"UTF-8\"");
     }
 
     /**
@@ -73,10 +111,15 @@ public class VirtuosoStorage extends Storage {
     private void makeVirtuosoPut(String uri, String data, String contentType) throws IOException {
         URL url = new URL(uri);
         HttpURLConnection conn = (HttpURLConnection) url.openConnection();
+        if (username != null && password != null) {
+            String auth = username + ":" + password;
+            String authHeaderValue = "Basic " + Base64.encode(auth.getBytes(CHARSET));
+            conn.setRequestProperty("Authorization", authHeaderValue);
+        }
         conn.setDoOutput(true);
         conn.setRequestProperty("Content-Type", contentType);
         conn.setRequestMethod("PUT");
-        OutputStreamWriter out = new OutputStreamWriter(conn.getOutputStream(), StandardCharsets.UTF_8);
+        OutputStreamWriter out = new OutputStreamWriter(conn.getOutputStream(), CHARSET);
         out.write(data);
         out.close();
         if (conn.getResponseCode() != HttpURLConnection.HTTP_CREATED) {