+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Possibility to specify custom storage in frontend
+
+### Fixed
+
+- Virtuoso storage can use HTTP Basic Auth
+
 ## [1.4.0]
 
 ### Changed